Implemented pardiso_64 linear solver and 64-bit version of CompactSymmMatrix.

+++ FECore/SparseMatrix.h
@@ -58,7 +58,7 @@ public:
 	bool IsSquare() const { return (m_nrow == m_ncol); }
 
 	//! return number of nonzeros
-	int NonZeroes() const { return m_nsize; }
+	size_t NonZeroes() const { return m_nsize; }
 
 public: // functions to be overwritten in derived classes
 
@@ -110,5 +110,5 @@ public:
 protected:
 	// NOTE: These values are set by derived classes
 	int	m_nrow, m_ncol;		//!< dimension of matrix
-	int	m_nsize;			//!< number of nonzeroes (i.e. matrix elements actually allocated)
+	size_t m_nsize;			//!< number of nonzeroes (i.e. matrix elements actually allocated)
 };
